api: make buffer public api
