+++ src/zarr/core/__init__.py
@@ -1,4 +1,3 @@
 from __future__ import annotations
 
-from zarr.core.buffer import Buffer, NDBuffer  # noqa: F401
 from zarr.core.codec_pipeline import BatchedCodecPipeline  # noqa: F401
