--- src/zarr/core/buffer/__init__.py
@@ -1,19 +1,0 @@
-from zarr.core.buffer.core import (
-    ArrayLike,
-    Buffer,
-    BufferPrototype,
-    NDArrayLike,
-    NDBuffer,
-    default_buffer_prototype,
-)
-from zarr.core.buffer.cpu import numpy_buffer_prototype
-
-__all__ = [
-    "ArrayLike",
-    "Buffer",
-    "BufferPrototype",
-    "NDArrayLike",
-    "NDBuffer",
-    "default_buffer_prototype",
-    "numpy_buffer_prototype",
-]
