+++ src/zarr/core/indexing.py
@@ -28,8 +28,8 @@ import numpy.typing as npt
 from zarr.core.common import product
 
 if TYPE_CHECKING:
+    from zarr.abc.buffer import NDArrayLike
     from zarr.core.array import Array
-    from zarr.core.buffer import NDArrayLike
     from zarr.core.chunk_grids import ChunkGrid
     from zarr.core.common import ChunkCoords
