+++ tests/test_array.py
@@ -12,6 +12,8 @@ import pytest
 
 import zarr.api.asynchronous
 from zarr import Array, AsyncArray, Group
+from zarr.buffer import default_buffer_prototype
+from zarr.buffer.cpu import NDBuffer
 from zarr.codecs import (
     BytesCodec,
     GzipCodec,
@@ -31,8 +33,6 @@ from zarr.core.array import (
     chunks_initialized,
     create_array,
 )
-from zarr.core.buffer import default_buffer_prototype
-from zarr.core.buffer.cpu import NDBuffer
 from zarr.core.chunk_grids import _auto_partition
 from zarr.core.common import JSON, MemoryOrder, ZarrFormat
 from zarr.core.group import AsyncGroup
