+++ tests/test_indexing.py
@@ -12,7 +12,7 @@ from numpy.testing import assert_array_equal
 
 import zarr
 from zarr import Array
-from zarr.core.buffer import default_buffer_prototype
+from zarr.buffer import default_buffer_prototype
 from zarr.core.indexing import (
     BasicSelection,
     CoordinateSelection,
@@ -31,8 +31,7 @@ from zarr.storage import MemoryStore, StorePath
 if TYPE_CHECKING:
     from collections.abc import AsyncGenerator
 
-    from zarr.core.buffer import BufferPrototype
-    from zarr.core.buffer.core import Buffer
+    from zarr.abc.buffer import Buffer, BufferPrototype
     from zarr.core.common import ChunkCoords
