+++ tests/test_metadata/test_v2.py
@@ -8,7 +8,7 @@ import pytest
 
 import zarr.api.asynchronous
 import zarr.storage
-from zarr.core.buffer import cpu
+from zarr.buffer import cpu
 from zarr.core.group import ConsolidatedMetadata, GroupMetadata
 from zarr.core.metadata import ArrayV2Metadata
 from zarr.core.metadata.v2 import parse_zarr_format
